# gbs-pos-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Point of Sale REST API for retail, F\u0026B, and fuel station operations",
        "title": "GBS POS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "GBS Support",
            "email": "support@gbs.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/v1/display/cart": {
            "get": {
                "description": "Returns the latest stored cart display JSON for a terminal. Public endpoint - no authentication required.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart Display"
                ],
                "summary": "Get cart display JSON",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Android terminal ID (ANDROID_ID)",
                        "name": "terminalId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON cart display data",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Receives cart display JSON from Android POS and stores it. Requires Bearer token authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Cart Display"
                ],
                "summary": "Upload cart display JSON",
                "parameters": [
                    {
                        "description": "Cart display JSON (must include terminalId field)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SaveCartDisplayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/v1/display/cart/{terminalId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Remove stored cart display for a terminal",
                "tags": [
                    "Cart Display"
                ],
                "summary": "Delete cart display",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Terminal ID",
                        "name": "terminalId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/v1/login": {
            "post": {
                "description": "Authenticate user with username and password. Returns JWT token for API access.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.LoginRequest": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.SaveCartDisplayRequest": {
            "type": "object",
            "required": [
                "terminalId"
            ],
            "properties": {
                "terminalId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token. For Keycloak, obtain token from {KEYCLOAK_BASE_URL}/realms/{KEYCLOAK_REALM}/protocol/openid-connect/token",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}